Reorganize functions into different files without meaningful code change
