+++ rtcrtpreceiver.go
@@ -1,0 +1,14 @@
+package webrtc
+
+// RTCRtpReceiver allows an application to inspect the receipt of a RTCTrack
+type RTCRtpReceiver struct {
+	Track *RTCTrack
+	// receiverTrack *RTCTrack
+	// receiverTransport
+	// receiverRtcpTransport
+}
+
+// TODO: receiving side
+// func newRTCRtpReceiver(kind, id string) {
+//
+// }
